separate demo folder
