--- chartlets.py/chartlets/demo/contribs/__init__.py
@@ -1,1 +1,0 @@
-from chartlets.demo.contribs.panel import Panel
