--- chartlets.py/chartlets/demo/utils/__init__.py
@@ -1,1 +1,0 @@
-from chartlets.demo.utils.json_encoder import NumpyJSONEncoder
