+++ chartlets.py/chartlets/version.py
@@ -1,1 +1,1 @@
-version = "0.0.30"
+version = "0.1.0.dev0"
