+++ chartlets.py/demo/server/contribs/__init__.py
@@ -1,0 +1,1 @@
+from .panel import Panel
