+++ chartlets.py/demo/server/contribs/panel.py
