+++ chartlets.py/demo/server/main.py
@@ -1,0 +1,11 @@
+import asyncio
+
+from demo.server.app import run_app
+
+
+def main():
+    asyncio.run(run_app())
+
+
+if __name__ == "__main__":
+    main()
